requirements.txt
flask
SQLALchemy
Flask-SQLAlchemy
Flask-MySQL 
PyMySQL
mysql-connector-python